Use the Kademlia distance metric, make servers redirect, ManyServersTest passes
Static factory methods for RawKey

+++ src/main/java/org/reefy/transportrest/api/Key.java
@@ -1,22 +1,22 @@
 package org.reefy.transportrest.api;
 
-import com.google.common.util.concurrent.ListenableFuture;
-
-import org.reefy.transportrest.api.store.Store;
-
-import java.nio.ByteBuffer;
+import java.math.BigInteger;
 
 /**
- * This represents a piece of data that can be stored on a single machine.
+ * This represents the identifier for a piece of data that can be stored on a single machine.
+ *
+ *
  *
  * @author Paul Kernfeld - pk@knewton.com
  */
 public interface Key {
     public byte[] getBytes();
 
-//    public ListenableFuture<V> get(Store store);
-//
-//    public void put(Store store, Value<V> testValue);
-
-    public int distance(Key key);
+    /**
+     * This should use Kademlia's XOR distance metric.
+     *
+     * @param key the other key to calculate the distance to.
+     * @return the distance between the two keys
+     */
+    public BigInteger distance(Key key);
 }
